Remove remaining sqrl/lib references from sqrl-redis-functions
