+++ packages/sqrl-cli/bin/cli.js
