+++ packages/sqrl-common/src/index.ts
@@ -1,16 +1,18 @@
-export * from "./compare";
-export * from "./ensureArray";
-export * from "./foreachObject";
-export * from "./hasConstructor";
-export * from "./flatten";
-export * from "./invariant";
-export * from "./jsonTemplate";
-export * from "./range";
-export * from "./isEmptyObject";
-export * from "./emptyFunction";
-export * from "./mapObject";
 export * from "./RenderedSpan";
 export * from "./SqrlObject";
+
+export * from "./compare";
+export * from "./emptyFunction";
+export * from "./ensureArray";
+export * from "./flatten";
+export * from "./foreachObject";
+export * from "./hasConstructor";
+export * from "./invariant";
+export * from "./isEmptyObject";
+export * from "./jsonTemplate";
+export * from "./mapObject";
+export * from "./promiseFinally";
+export * from "./range";
 export * from "./sqrlCartesianProduct";
 
 export interface AssertService {
