+++ packages/sqrl-test-utils/package.json
@@ -2,7 +2,6 @@
   "name": "sqrl-test-utils",
   "version": "0.5.0",
   "description": "> TODO: description",
-  "homepage": "",
   "directories": {
     "lib": "lib",
     "test": "__tests__"
