+++ packages/sqrl/__tests__/helpers/runCompile.ts
@@ -12,9 +12,9 @@ import { statementsFromString } from "../../src/helpers/CompileHelpers";
 import { JsExecutionContext } from "../../src/execute/JsExecutionContext";
 import SqrlExecutable from "../../src/execute/SqrlExecutable";
 import { createDefaultContext } from "../../src/helpers/ContextHelpers";
-import { SimpleManipulator } from "../../src/simple/SimpleManipulator";
 import { Filesystem } from "../../src/api/filesystem";
 import { FeatureMap } from "../../src/api/execute";
+import { SimpleManipulator } from "../../src/api/simple/SimpleManipulator";
 
 export class VirtualSourceTree extends Filesystem {
   constructor(
