+++ packages/sqrl/__tests__/helpers/sqrlTest.ts
@@ -1,4 +1,4 @@
-import { runSqrlTest } from "../../src/testing/runSqrlTest";
+import { runSqrlTest } from "../../src/api/simple/runSqrlTest";
 
 export function sqrlTest(name: string, sqrl: string) {
   test(name, () => runSqrlTest(sqrl));
