+++ packages/sqrl/__tests__/sqrl/lazy.spec.ts
@@ -3,13 +3,11 @@
  * Licensed under the Apache License, Version 2.0
  * http://www.apache.org/licenses/LICENSE-2.0
  */
-import {
-  runSqrlTest,
-  buildTestFunctionRegistry
-} from "../../src/testing/runSqrlTest";
+import { buildTestFunctionRegistry } from "../../src/testing/runSqrlTest";
 import { FunctionCostData } from "../../src/function/FunctionRegistry";
 import { registerTestFunctions } from "../helpers/TestFunctions";
 import { FunctionRegistry } from "../../src/api/execute";
+import { runSqrlTest } from "../../src/api/simple/runSqrlTest";
 
 test("works", async () => {
   let functionRegistry: FunctionRegistry;
