+++ packages/sqrl/src/index.ts
@@ -15,6 +15,7 @@ export * from "./api/object";
 export * from "./api/parse";
 export * from "./api/test";
 export * from "./api/when";
+export * from "./api/simple/runSqrlTest";
 
 export * from "./api/ArgumentCheck";
 export * from "./api/AstBuilder";
@@ -22,5 +23,9 @@ export * from "./api/ExecutableSpec";
 export * from "./api/Manipulator";
 export * from "./api/UniqueId";
 export * from "./api/UniqueIdService";
+export * from "./api/simple/SimpleBlockService";
+export * from "./api/simple/SimpleLogService";
+export * from "./api/simple/SimpleManipulator";
+export * from "./api/simple/TestLogger";
 
 export { AssertService, sqrlCompare } from "sqrl-common";
